refactor(sync): rename sync-core → sync, sync → sync-client
The protocol package deserves the clean name since it's the shared
foundation. The client provider is the specialized consumer.
