+++ docs/articles/y-sweet-message-sync-status.md
@@ -1,6 +1,6 @@
 # y-sweet's MESSAGE_SYNC_STATUS: Application-Layer TCP Sequence Numbers
 
-> **Note**: `@epicenter/sync-core` and `@epicenter/sync` implement this protocol. See `packages/sync-core/src/protocol.ts` for the server-side encoding and `packages/sync/src/provider.ts` for the client-side implementation. The article below describes y-sweet's original implementation, which inspired ours.
+> **Note**: `@epicenter/sync` and `@epicenter/sync` implement this protocol. See `packages/sync-core/src/protocol.ts` for the server-side encoding and `packages/sync/src/provider.ts` for the client-side implementation. The article below describes y-sweet's original implementation, which inspired ours.
 
 Yjs's sync protocol is fire-and-forget — you serialize your local changes into a binary update, shove it into the WebSocket, and hope for the best. There's no built-in acknowledgment. The problem is that WebSockets can silently die (the TCP connection hangs, the server crashes, your laptop switches from WiFi to cellular) and your application has no way to know that the last 30 seconds of typing never left the machine. The local version counter fixes this cheaply: every time you make a local edit, you bump an integer and send it alongside. The server echoes it back. Now you can compare two numbers — `ackedVersion` vs `localVersion` — and know instantly whether your changes have reached the server. This gives you three things you couldn't have otherwise: (1) a reliable `hasLocalChanges` flag so you can show a "syncing..." indicator or block the `beforeunload` with "you have unsaved changes," (2) a liveness probe that doubles as the heartbeat — if the echo doesn't come back in 3 seconds, you know the connection is dead and can reconnect *before* the user notices, and (3) you get all of this without touching the Yjs document state at all — the server doesn't even parse the version number, it just mirrors bytes, so it's zero-cost and can never corrupt sync state. It's essentially the same idea as TCP sequence numbers but at the application layer: tag your outbound data with a monotonic counter, wait for the other side to confirm it, and if confirmation stops arriving, assume the pipe is broken.
