+++ packages/server-remote-cloudflare/package.json
@@ -19,7 +19,7 @@
 	"dependencies": {
 		"@better-auth/drizzle-adapter": "catalog:",
 		"@better-auth/oauth-provider": "catalog:",
-		"@epicenter/sync-core": "workspace:*",
+		"@epicenter/sync": "workspace:*",
 		"@hono/standard-validator": "^0.2.2",
 		"@tanstack/ai": "^0.5.1",
 		"@tanstack/ai-anthropic": "^0.5.0",
