+++ packages/sync-client/src/index.ts
@@ -1,0 +1,14 @@
+export type {
+	HttpSyncProvider,
+	HttpSyncProviderConfig,
+	HttpSyncStatus,
+} from './http-provider';
+export { createHttpSyncProvider } from './http-provider';
+export { createSyncProvider } from './provider';
+export { createSleeper, type Sleeper } from './sleeper';
+export type {
+	SyncProvider,
+	SyncProviderConfig,
+	SyncStatus,
+	WebSocketConstructor,
+} from './types';
