+++ packages/sync-client/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"lib": ["ES2022", "DOM"]
+	},
+	"include": ["src/**/*"]
+}
