--- packages/sync-core/tsconfig.json
@@ -1,29 +1,0 @@
-{
-	"compilerOptions": {
-		"lib": ["ESNext", "DOM"],
-		"target": "es2022",
-		"module": "Preserve",
-		"moduleDetection": "force",
-		"allowJs": true,
-		"resolveJsonModule": true,
-		"esModuleInterop": true,
-		"isolatedModules": true,
-
-		"moduleResolution": "bundler",
-		"allowImportingTsExtensions": true,
-		"verbatimModuleSyntax": true,
-		"noEmit": true,
-
-		"sourceMap": true,
-
-		"strict": true,
-		"skipLibCheck": true,
-		"noFallthroughCasesInSwitch": true,
-		"noUncheckedIndexedAccess": true,
-		"noImplicitOverride": true,
-
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
-		"noPropertyAccessFromIndexSignature": false
-	}
-}
