+++ packages/sync/package.json
@@ -1,19 +1,24 @@
 {
 	"name": "@epicenter/sync",
-	"version": "0.1.0",
-	"description": "Yjs WebSocket sync provider with supervisor loop architecture.",
-	"license": "AGPL-3.0",
+	"version": "0.0.1",
 	"main": "./src/index.ts",
 	"types": "./src/index.ts",
 	"exports": {
 		".": "./src/index.ts"
 	},
+	"license": "AGPL-3.0",
+	"scripts": {
+		"typecheck": "tsc --noEmit"
+	},
 	"dependencies": {
-		"@epicenter/sync-core": "workspace:*",
 		"lib0": "catalog:",
 		"y-protocols": "catalog:"
 	},
 	"peerDependencies": {
 		"yjs": "catalog:"
+	},
+	"devDependencies": {
+		"@types/bun": "catalog:",
+		"typescript": "catalog:"
 	}
 }
