+++ packages/sync/tsconfig.json
@@ -1,7 +1,29 @@
 {
-	"extends": "../../tsconfig.base.json",
 	"compilerOptions": {
-		"lib": ["ES2022", "DOM"]
-	},
-	"include": ["src/**/*"]
+		"lib": ["ESNext", "DOM"],
+		"target": "es2022",
+		"module": "Preserve",
+		"moduleDetection": "force",
+		"allowJs": true,
+		"resolveJsonModule": true,
+		"esModuleInterop": true,
+		"isolatedModules": true,
+
+		"moduleResolution": "bundler",
+		"allowImportingTsExtensions": true,
+		"verbatimModuleSyntax": true,
+		"noEmit": true,
+
+		"sourceMap": true,
+
+		"strict": true,
+		"skipLibCheck": true,
+		"noFallthroughCasesInSwitch": true,
+		"noUncheckedIndexedAccess": true,
+		"noImplicitOverride": true,
+
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"noPropertyAccessFromIndexSignature": false
+	}
 }
